+++ pdtable/test/io/test__excel.py
@@ -10,7 +10,7 @@ except ImportError:
     from openpyxl.worksheet import Worksheet as OpenpyxlWorksheet
 
 from pdtable import Table
-from pdtable.io._excel import write_excel
+from pdtable.io.excel import write_excel
 from pdtable.io._excel_openpyxl import _append_table_to_openpyxl_worksheet
